initial push to create test coverage for the attach debug dialog
